mcp_server/requirements.txt
# Fast Info MCP Server Dependencies
mcp>=1.0.0
fastmcp>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Testing
pytest>=8.0.0
pytest-mock>=3.12.0
